🎬 :: move solution class to model
